move_trace_api2
